+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Fixed table row height in pokédex while waiting for images to load.
 
+- Installed Django debug toolbar
+
+- Fixed an issue where pokedexes with owners could not be accessed.
+
 ## 12/3-2023
 
 - Pokédex model in the admin page no longer crashes at it attempts to load all entries.